feat : 팔로워 조회 API에 페이지네이션 적용

+++ pickly-service/src/main/java/org/pickly/service/bookmark/repository/impl/BookmarkQueryRepositoryImpl.java
@@ -1,16 +1,17 @@
 package org.pickly.service.bookmark.repository.impl;
 
-import static org.pickly.service.bookmark.entity.QBookmark.bookmark;
-
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.pickly.service.bookmark.entity.Bookmark;
 import org.pickly.service.bookmark.entity.Visibility;
 import org.pickly.service.bookmark.repository.interfaces.BookmarkQueryRepository;
 import org.pickly.service.common.utils.page.PageRequest;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+import static org.pickly.service.bookmark.entity.QBookmark.bookmark;
 
 @Repository
 @RequiredArgsConstructor
@@ -25,7 +26,7 @@ public class BookmarkQueryRepositoryImpl implements BookmarkQueryRepository {
       PageRequest pageRequest, Long memberId, Long categoryId, Boolean isUserLike,
       Boolean readByUser, Visibility visibility
   ) {
-    Long cursorId = pageRequest.getCursorId();
+    Long cursorId = (Long) pageRequest.getCursorId();
     Integer pageSize = pageRequest.getPageSize();
     return queryFactory
         .selectFrom(bookmark)
@@ -57,7 +58,7 @@ public class BookmarkQueryRepositoryImpl implements BookmarkQueryRepository {
   @Override
   public List<Bookmark> findBookmarkByCategoryId(PageRequest pageRequest, Long categoryId) {
 
-    Long cursorId = pageRequest.getCursorId();
+    Long cursorId = (Long) pageRequest.getCursorId();
     Integer pageSize = pageRequest.getPageSize();
 
     return queryFactory
